quality/requirements.txt
flask==1.1.1
jira==2.0.0
pymysql==0.9.3
requests==2.22.0
qrcode==6.1
pillow==6.2.0
gunicorn==20.0.4
gevent==1.4.0
python-gitlab==1.13.0
Flask-Cors==3.0.8
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-Script==2.0.6
flask_sqlalchemy==2.4.1
flask-login==0.4.1
Jinja2==2.10.3
mysql-connector==2.2.9
python-ldap==3.2.0
casbin==0.8.2
casbin-sqlalchemy-adapter==0.1.0
python-ldap==3.2.0
SQLAlchemy==1.3.5
Flask-APScheduler==1.11.0
kafka-python>=2.0.0
